Remove q-arm check as assumptions too unreliable.
