+++ censtats/status/reference.py
@@ -1,8 +1,7 @@
 import re
 import polars as pl
 from typing import NamedTuple, Generator
-from .acrocentrics import get_q_arm_acro_chr
-from .constants import ACROCENTRIC_CHROMOSOMES, RGX_CHR
+from .constants import RGX_CHR
 
 
 class RefCenContigs(NamedTuple):
@@ -22,9 +21,4 @@ def split_ref_rm_input_by_contig(
 
         ref_chr_name = mtch_ref_chr_name.group()
 
-        # Also adjust for reference acrocentrics.
-        if ref_chr_name in ACROCENTRIC_CHROMOSOMES:
-            df_ref_qarm_grp = get_q_arm_acro_chr(df_ref_grp)
-            yield ref, RefCenContigs(ref_chr_name, ref, df_ref_qarm_grp)
-        else:
-            yield ref, RefCenContigs(ref_chr_name, ref, df_ref_grp)
+        yield ref, RefCenContigs(ref_chr_name, ref, df_ref_grp)
