+++ pyproject.toml
@@ -15,7 +15,7 @@ readme = "README.md"
 requires-python = ">=3.12"
 license = {text = "MIT License"}
 dynamic = ["dependencies"]
-version = "0.0.4"
+version = "0.0.5"
 
 [project.urls]
 Homepage = "https://github.com/logsdon-lab/CenStats"
