+++ test/status/expected/correct_chr21_cens.tsv
@@ -1,4 +1,4 @@
 HG00171_chr21_haplotype1-0000029:4196961-10792258	HG00171_chr13_haplotype1-0000029:4196961-10792258	fwd	false
-HG00171_chr21_haplotype1-0000019:33327260-37313906	HG00171_chr9_haplotype1-0000019:33327260-37313906	fwd	true
+HG00171_chr21_haplotype1-0000019:33327260-37313906	HG00171_chr21_haplotype1-0000019:33327260-37313906	rev	true
 HG00171_chr21_haplotype2-0000154:33297526-38584922	HG00171_chr21_haplotype2-0000154:33297526-38584922	rev	false
 chm13_chr21:7700001-11850000	chm13_chr21:7700001-11850000	fwd	false
