+++ test/status/expected/correct_chr21_cens_false_neg_mismap.tsv
@@ -1,1 +1,1 @@
-NA24385_chr21_h2tg000030l#1-40398587:38371238-40398587	NA24385_chr13_h2tg000030l#1-40398587:38371238-40398587	rev	false
+NA24385_chr21_h2tg000030l#1-40398587:38371238-40398587	NA24385_chr21_h2tg000030l#1-40398587:38371238-40398587	rev	false
