+++ test/status/expected/correct_chr21_chr13_cens_mismap.tsv
@@ -1,3 +1,3 @@
-correct_chr21_unassigned:2-2293201	correct_chr21_unassigned:2-2293201	rev	false
+correct_chr21_unassigned:2-2293201	correct_chr21_unassigned:2-2293201	fwd	false
 mismapped_chr21_unassigned:2078337-9206433	mismapped_chr13_unassigned:2078337-9206433	rev	false
 chm13_chr21:7700001-11850000	chm13_chr21:7700001-11850000	fwd	false
